Renamed gem to apcera-stager-api
