+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 All notable changes to this project will be documented in this file.
+Documentation started at the 0.2.5 release.
 
-## 0.0.1 Initial version
+## 0.2.5 apcera-stager-api
 
 ### Added
 - Nothing.
@@ -13,4 +14,4 @@ All notable changes to this project will be documented in this file.
 - Nothing.
 
 ### Fixed
-- Nothing.
+- Has been renamed to apcera-stager-api
