+++ Rakefile
@@ -1,4 +1,5 @@
 #!/usr/bin/env rake
+
 require File.join('bundler', 'gem_tasks')
 require File.join('rspec', 'core', 'rake_task')
