+++ apcera-stager-api-contrib.gemspec
@@ -5,17 +5,17 @@ Gem::Specification.new do |gem|
 
   gem.authors       = ["Josh Ellithorpe"]
   gem.email         = ["josh@apcera.com"]
-  gem.description   = %q{Continuum Stager api library}
-  gem.summary       = %q{Continuum Stager api library which makes it super easy to write stagers for Apcera's Continuum.}
+  gem.description   = %q{Apcera Stager API Library}
+  gem.summary       = %q{Apcera Stager API library which makes it super easy to write stagers for the Apcera Platform.}
   gem.homepage      = "http://apcera.com"
   gem.license       = "MIT"
 
   gem.executables   = `git ls-files -- bin/*`.split("\n").map{ |f| File.basename(f) }
   gem.files         = `git ls-files`.split("\n")
   gem.test_files    = `git ls-files -- {spec}/*`.split("\n")
-  gem.name          = "continuum-stager-api"
+  gem.name          = "apcera-stager-api"
   gem.require_paths = ["lib"]
-  gem.version       = "0.2.4"
+  gem.version       = "0.2.5"
 
   gem.add_development_dependency 'rspec', '~> 2.6.0'
   gem.add_development_dependency 'rake'
