+++ lib/apcera-stager-api.rb
